Initial commit.
To do: I think the Forces of Demand visualizations are week and have
significant room for improvement. Also, does System of Progress belong
in a 101 presentation? I think it’s helpful for understanding the
nature of JTBD. Finally, it seems like this would benefit from the
inclusion of some examples like the Snickers or Bob’s homebuilding
story.